+++ packages/coding-standards/index.js
@@ -1,5 +1,4 @@
 module.exports = {
-	commitlintConfig: require("./commitlint"),
 	cspellConfig: require("./cspell/index.cjs"),
 	prettierConfig: require("./prettier/index.cjs"),
 	stylelintConfig: require("./stylelint/index.cjs"),