+++ packages/compiler/src/index.ts
@@ -1,8 +1,2 @@
-import type { Options } from "sass";
 // export { config } from "./config";
 // export { config as default } from "./config";
-
-export type SCSSAliases = {
-	loadPaths?: Options<"async">["loadPaths"];
-	importers?: Options<"async">["importers"];
-};