+++ packages/compiler/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @atomicsmash/compiler
 
+## 3.0.0-beta.3
+
+### Major Changes
+
+- [#480](https://github.com/AtomicSmash/packages/pull/480) [`9e83ae4`](https://github.com/AtomicSmash/packages/commit/9e83ae42689cc7cea608db2414337d676d1ff083) Thanks [@mikeybinns](https://github.com/mikeybinns)! - Use smash config package for loading smash config
+
+### Patch Changes
+
+- Updated dependencies [[`9e83ae4`](https://github.com/AtomicSmash/packages/commit/9e83ae42689cc7cea608db2414337d676d1ff083)]:
+  - @atomicsmash/smash-config@1.0.0-beta.0
+
+## 3.0.0-beta.2
+
+### Patch Changes
+
+- Updated dependencies [[`28787b5`](https://github.com/AtomicSmash/packages/commit/28787b5ab36f01c92ab7f9f899a2fdc201614fe0), [`ef8fab3`](https://github.com/AtomicSmash/packages/commit/ef8fab35fdd40d14096439d3d35d97e7c1addab0)]:
+  - @atomicsmash/cli@10.0.0-beta.1
+
+## 3.0.0-beta.1
+
+### Patch Changes
+
+- [#476](https://github.com/AtomicSmash/packages/pull/476) [`02f62d2`](https://github.com/AtomicSmash/packages/commit/02f62d24798dd0e6931d062c3b168adeceee54b4) Thanks [@mikeybinns](https://github.com/mikeybinns)! - Add PHP files to the file dependencies array to trigger watch rebuilds
+
+## 3.0.0-beta.0
+
+### Patch Changes
+
+- Updated dependencies [[`197e841`](https://github.com/AtomicSmash/packages/commit/197e8410cb516916a76f6aba51ed5782052064b6)]:
+  - @atomicsmash/cli@10.0.0-beta.0
+
 ## 2.0.1
 
 ### Patch Changes