+++ packages/wordpress-tests-helper/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @atomicsmash/wordpress-tests-helper
 
+## 1.2.0-beta.1
+
+### Patch Changes
+
+- [#485](https://github.com/AtomicSmash/packages/pull/485) [`3479e4d`](https://github.com/AtomicSmash/packages/commit/3479e4d5d03b82daaa347173aeb81aa2161207c6) Thanks [@mikeybinns](https://github.com/mikeybinns)! - Fix older WP versions with 6.9 changes
+
+## 1.2.0-beta.0
+
+### Minor Changes
+
+- [#483](https://github.com/AtomicSmash/packages/pull/483) [`87ab88a`](https://github.com/AtomicSmash/packages/commit/87ab88a5812a903372eb8868d6a174d974f3fe12) Thanks [@mikeybinns](https://github.com/mikeybinns)! - Add support for WP 6.9
+
 ## 1.1.0
 
 ### Minor Changes